Move abm.cpp to simulations folder and first attemp to run abm simulation multiple times
