+++ cpp/examples/CMakeLists.txt
@@ -26,9 +26,6 @@ target_link_libraries(ode_secir_ageres_example PRIVATE memilio ode_secir)
 add_executable(graph_example graph.cpp)
 target_link_libraries(graph_example PRIVATE memilio ode_seir)
 
-add_executable(abm_example abm.cpp)
-target_link_libraries(abm_example PRIVATE memilio abm)
-
 add_executable(abm_minimal_example abm_minimal.cpp)
 target_link_libraries(abm_minimal_example PRIVATE memilio abm)
